Refactor integrated tests
