+++ Web-Api.Tests/Controllers/AuthorControllerTest.cs
@@ -1,10 +1,10 @@
-using System.Net;
-using System.Net.Http.Json;
-using ApiBookStore;
-using BLL.DTO.Book;
-using DLL.Models;
+﻿using ApiBookStore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using System.Net;
+using System.Net.Http.Json;
+using BLL.DTO.Author;
+using DLL.Models;
 using Newtonsoft.Json;
 using Web_Api.Tests.Extensions;
 using Web_Api.Tests.Startup;
@@ -13,24 +13,24 @@ using Xunit;
 
 namespace Web_Api.Tests.Controllers
 {
-    public class BookControllerTest : IClassFixture<WebApplicationFactoryTest<Program>>
+    public class AuthorControllerTest : IClassFixture<WebApplicationFactoryTest<Program>>
     {
         private readonly WebApplicationFactoryTest<Program> _appFactory;
         private readonly TokenServiceTest _tokenJwtService;
 
-        public BookControllerTest(WebApplicationFactoryTest<Program> appFactory)
+        public AuthorControllerTest(WebApplicationFactoryTest<Program> appFactory)
         {
             _appFactory = appFactory;
 
             using var configScope = _appFactory.Services.CreateScope();
             var config = configScope.ServiceProvider.GetRequiredService<IConfiguration>();
-            
+
             _tokenJwtService = new TokenServiceTest(config);
         }
 
         [Theory]
-        [InlineData("/api/book")]
-        public async Task BookGetAllAsyncTask_Return_Ok(string url)
+        [InlineData("/api/author")]
+        public async Task AuthorGetAllAsyncTask_Return_Ok(string url)
         {
             // Arranges
             var client = _appFactory.CreateClient();
@@ -44,8 +44,9 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("/api/book/1")]
-        public async Task BookGetByIdAsyncTask_Return_Ok(string url)
+        [InlineData("/api/author/1")]
+        [InlineData("/api/author/2")]
+        public async Task AuthorGetByIdAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -59,42 +60,38 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Created_201(string url)
+        [InlineData("api/author")]
+        public async Task AuthorCreateAsyncTask_Return_Created_201(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
-            
+
             var tokenJwt = _tokenJwtService.CreateTokenManagerRole("Manager");
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var createBookDto = new CreateBookDto
+            var createAuthorDto = new CreateAuthorDto
             {
-                Name = "Test-book",
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                FirstName = "Test-first-name",
+                LastName = "Test-last-name"
             };
 
             // Act
-            var response = await client.PostAsJsonAsync(url, createBookDto);
+            var response = await client.PostAsJsonAsync(url, createAuthorDto);
 
-            var bookString = await response.Content.ReadAsStringAsync();
-            var bookCreated = JsonConvert.DeserializeObject<Book>(bookString)!;
+            var authorString = await response.Content.ReadAsStringAsync();
+            var authorCreated = JsonConvert.DeserializeObject<Author>(authorString)!;
 
             // Assert
             Assert.Equal(HttpStatusCode.Created, response.StatusCode);
-            Assert.Equal(createBookDto.Name, bookCreated.Name);
-            Assert.Equal(createBookDto.Price, bookCreated.Price);
-            Assert.Equal(createBookDto.IdPublisher, bookCreated.Publisher.Id);
+            Assert.Equal(createAuthorDto.FirstName, authorCreated.FirstName);
+            Assert.Equal(createAuthorDto.LastName, authorCreated.LastName);
             Assert.Equal("application/json; charset=utf-8", response.Content.Headers.ContentType?.ToString());
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_BadRequest_400(string url)
+        [InlineData("api/author")]
+        public async Task AuthorCreateAsyncTask_Return_BadRequest_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -104,17 +101,17 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            CreateBookDto bookDtoNull = null;
+            CreateAuthorDto authorDtoNull = null;
             // ReSharper disable once ExpressionIsAlwaysNull
-            var response = await client.PostAsJsonAsync(url, bookDtoNull);
+            var response = await client.PostAsJsonAsync(url, authorDtoNull);
 
             // Assert
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/author")]
+        public async Task AuthorCreateAsyncTask_Return_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -124,13 +121,10 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreateAuthorDto
             {
-                Name = "Test-book",
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                FirstName = "Test-first-name",
+                LastName = "Test-last-name"
             });
 
             // Assert
@@ -138,8 +132,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_ValidationError_400(string url)
+        [InlineData("api/author")]
+        public async Task AuthorCreateAsyncTask_Return_ValidationError_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -149,13 +143,10 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreateAuthorDto
             {
-                Name = string.Empty,
-                Price = -99.99m,
-                IdPublisher = 0,
-                GenresId = new List<int> { 0 },
-                AuthorsId = new List<int> { 0 }
+                FirstName = string.Empty,
+                LastName = string.Empty
             });
 
             // Assert
@@ -163,19 +154,17 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/author")]
+        public async Task AuthorCreateAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreateAuthorDto
             {
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                FirstName = "Test-first-name",
+                LastName = "Test-last-name"
             });
 
             // Assert
@@ -183,8 +172,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Ok(string url)
+        [InlineData("api/author")]
+        public async Task AuthorUpdateAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -193,34 +182,30 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updateAuthorDto = new UpdateAuthorDto()
             {
-                Id = 2,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Id = 1,
+                FirstName = "Test-first-name-EDIT",
+                LastName = "Test-last-name-EDIT"
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updateAuthorDto);
 
-            var bookString = await response.Content.ReadAsStringAsync();
-            var bookUpdated = JsonConvert.DeserializeObject<Book>(bookString);
+            var authorString = await response.Content.ReadAsStringAsync();
+            var authorUpdated = JsonConvert.DeserializeObject<Author>(authorString);
 
             // Assert
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            Assert.Equal(updateBookDto.Id, bookUpdated!.Id);
-            Assert.Equal(updateBookDto.Name, bookUpdated.Name);
-            Assert.Equal(updateBookDto.Price, bookUpdated.Price);
-            Assert.Equal(updateBookDto.IdPublisher, bookUpdated.Publisher.Id);
+            Assert.Equal(updateAuthorDto.Id, authorUpdated!.Id);
+            Assert.Equal(updateAuthorDto.FirstName, authorUpdated.FirstName);
+            Assert.Equal(updateAuthorDto.LastName, authorUpdated.LastName);
             Assert.Equal("application/json; charset=utf-8", response.Content.Headers.ContentType?.ToString());
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/author")]
+        public async Task AuthorUpdateAsyncTask_Return_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -229,50 +214,44 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updateAuthorDto = new UpdateAuthorDto()
             {
                 Id = 1,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                FirstName = "Test-first-name-EDIT",
+                LastName = "Test-last-name-EDIT"
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updateAuthorDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.Forbidden, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/author")]
+        public async Task AuthorUpdateAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
 
-            var updateBookDto = new UpdateBookDto()
+            var updateAuthorDto = new UpdateAuthorDto()
             {
                 Id = 1,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                FirstName = "Test-first-name-EDIT",
+                LastName = "Test-last-name-EDIT"
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
-            
+            var response = await client.PutAsJsonAsync(url, updateAuthorDto);
+
             // Assert
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_ValidationError_400(string url)
+        [InlineData("api/author")]
+        public async Task AuthorUpdateAsyncTask_Return_ValidationError_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -281,26 +260,23 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updateAuthorDto = new UpdateAuthorDto()
             {
                 Id = 1,
-                Name = string.Empty,
-                Price = -99.99m,
-                IdPublisher = 0,
-                GenresId = new List<int> { 0 },
-                AuthorsId = new List<int> { 0 }
+                FirstName = string.Empty,
+                LastName = "Test-last-name-EDIT"
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updateAuthorDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_NotFound_404(string url)
+        [InlineData("api/author")]
+        public async Task AuthorUpdateAsyncTask_Return_NotFound_404(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -309,26 +285,23 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updateAuthorDto = new UpdateAuthorDto()
             {
-                Id = 99999,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Id = 999999999,
+                FirstName = "Test-first-name",
+                LastName = "Test-last-name"
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updateAuthorDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book/2")]
-        public async Task BookDeleteAsyncTask_ReturnOk(string url)
+        [InlineData("api/author/3")]
+        public async Task AuthorDeleteAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -345,8 +318,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book/3")]
-        public async Task BookDeleteAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/author/3")]
+        public async Task AuthorDeleteAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -359,8 +332,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book/2")]
-        public async Task BookDeleteAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/author/2")]
+        public async Task AuthorDeleteAsyncTask_Return_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
