+++ Web-Api.Tests/Controllers/PublisherControllerTest.cs
@@ -1,11 +1,11 @@
-using System.Net;
-using System.Net.Http.Json;
-using ApiBookStore;
-using BLL.DTO.Book;
-using DLL.Models;
+﻿using ApiBookStore;
+using BLL.DTO.Publisher;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
+using System.Net;
+using System.Net.Http.Json;
+using DLL.Models;
 using Web_Api.Tests.Extensions;
 using Web_Api.Tests.Startup;
 using Web_Api.Tests.Startup.JwtHandler;
@@ -13,24 +13,24 @@ using Xunit;
 
 namespace Web_Api.Tests.Controllers
 {
-    public class BookControllerTest : IClassFixture<WebApplicationFactoryTest<Program>>
+    public class PublisherControllerTest : IClassFixture<WebApplicationFactoryTest<Program>>
     {
         private readonly WebApplicationFactoryTest<Program> _appFactory;
         private readonly TokenServiceTest _tokenJwtService;
 
-        public BookControllerTest(WebApplicationFactoryTest<Program> appFactory)
+        public PublisherControllerTest(WebApplicationFactoryTest<Program> appFactory)
         {
             _appFactory = appFactory;
 
             using var configScope = _appFactory.Services.CreateScope();
             var config = configScope.ServiceProvider.GetRequiredService<IConfiguration>();
-            
+
             _tokenJwtService = new TokenServiceTest(config);
         }
 
         [Theory]
-        [InlineData("/api/book")]
-        public async Task BookGetAllAsyncTask_Return_Ok(string url)
+        [InlineData("/api/publisher")]
+        public async Task PublisherGetAllAsyncTask_Return_Ok(string url)
         {
             // Arranges
             var client = _appFactory.CreateClient();
@@ -44,8 +44,9 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("/api/book/1")]
-        public async Task BookGetByIdAsyncTask_Return_Ok(string url)
+        [InlineData("/api/publisher/1")]
+        [InlineData("/api/publisher/2")]
+        public async Task PublisherGetByIdAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -59,42 +60,36 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Created_201(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherCreateAsyncTask_Return_Created_201(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
-            
+
             var tokenJwt = _tokenJwtService.CreateTokenManagerRole("Manager");
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var createBookDto = new CreateBookDto
+            var createPublisherDto = new CreatePublisherDto
             {
-                Name = "Test-book",
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Name = "Test-publisher-name"
             };
 
             // Act
-            var response = await client.PostAsJsonAsync(url, createBookDto);
+            var response = await client.PostAsJsonAsync(url, createPublisherDto);
 
-            var bookString = await response.Content.ReadAsStringAsync();
-            var bookCreated = JsonConvert.DeserializeObject<Book>(bookString)!;
+            var publisherString = await response.Content.ReadAsStringAsync();
+            var publisherCreated = JsonConvert.DeserializeObject<Publisher>(publisherString)!;
 
             // Assert
             Assert.Equal(HttpStatusCode.Created, response.StatusCode);
-            Assert.Equal(createBookDto.Name, bookCreated.Name);
-            Assert.Equal(createBookDto.Price, bookCreated.Price);
-            Assert.Equal(createBookDto.IdPublisher, bookCreated.Publisher.Id);
+            Assert.Equal(createPublisherDto.Name, publisherCreated.Name);
             Assert.Equal("application/json; charset=utf-8", response.Content.Headers.ContentType?.ToString());
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_BadRequest_400(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherCreateAsyncTask_Return_BadRequest_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -104,17 +99,17 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            CreateBookDto bookDtoNull = null;
+            CreatePublisherDto publisherDtoNull = null;
             // ReSharper disable once ExpressionIsAlwaysNull
-            var response = await client.PostAsJsonAsync(url, bookDtoNull);
+            var response = await client.PostAsJsonAsync(url, publisherDtoNull);
 
             // Assert
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherCreateAsyncTask_Return_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -124,13 +119,9 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreatePublisherDto
             {
-                Name = "Test-book",
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Name = "Test-publisher-name",
             });
 
             // Assert
@@ -138,8 +129,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_ValidationError_400(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherCreateAsyncTask_Return_ValidationError_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -149,13 +140,9 @@ namespace Web_Api.Tests.Controllers
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreatePublisherDto
             {
                 Name = string.Empty,
-                Price = -99.99m,
-                IdPublisher = 0,
-                GenresId = new List<int> { 0 },
-                AuthorsId = new List<int> { 0 }
             });
 
             // Assert
@@ -163,19 +150,16 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookCreateAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherCreateAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
 
             // Act
-            var response = await client.PostAsJsonAsync(url, new CreateBookDto
+            var response = await client.PostAsJsonAsync(url, new CreatePublisherDto
             {
-                Price = 99.99m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Name = "Test-publisher-name",
             });
 
             // Assert
@@ -183,8 +167,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Ok(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherUpdateAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -193,34 +177,28 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updatePublisherDto = new UpdatePublisherDto()
             {
                 Id = 2,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Name = "Test-publisher-name-EDIT",
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updatePublisherDto);
 
-            var bookString = await response.Content.ReadAsStringAsync();
-            var bookUpdated = JsonConvert.DeserializeObject<Book>(bookString);
+            var publisherString = await response.Content.ReadAsStringAsync();
+            var publisherUpdated = JsonConvert.DeserializeObject<Publisher>(publisherString);
 
             // Assert
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
-            Assert.Equal(updateBookDto.Id, bookUpdated!.Id);
-            Assert.Equal(updateBookDto.Name, bookUpdated.Name);
-            Assert.Equal(updateBookDto.Price, bookUpdated.Price);
-            Assert.Equal(updateBookDto.IdPublisher, bookUpdated.Publisher.Id);
+            Assert.Equal(updatePublisherDto.Id, publisherUpdated!.Id);
+            Assert.Equal(updatePublisherDto.Name, publisherUpdated.Name);
             Assert.Equal("application/json; charset=utf-8", response.Content.Headers.ContentType?.ToString());
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherUpdateAsyncTask_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -229,50 +207,42 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updatePublisherDto = new UpdatePublisherDto()
             {
-                Id = 1,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Id = 3,
+                Name = "Test-publisher-name-EDIT",
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updatePublisherDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.Forbidden, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherUpdateAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
 
-            var updateBookDto = new UpdateBookDto()
+            var updatePublisherDto = new UpdatePublisherDto()
             {
                 Id = 1,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Name = "Test-publisher-name-EDIT",
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
-            
+            var response = await client.PutAsJsonAsync(url, updatePublisherDto);
+
             // Assert
             Assert.Equal(HttpStatusCode.Unauthorized, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_ValidationError_400(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherUpdateAsyncTask_Return_ValidationError_400(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -281,26 +251,22 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updatePublisherDto = new UpdatePublisherDto()
             {
                 Id = 1,
                 Name = string.Empty,
-                Price = -99.99m,
-                IdPublisher = 0,
-                GenresId = new List<int> { 0 },
-                AuthorsId = new List<int> { 0 }
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updatePublisherDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book")]
-        public async Task BookUpdateAsyncTask_Return_NotFound_404(string url)
+        [InlineData("api/publisher")]
+        public async Task PublisherUpdateAsyncTask_Return_NotFound_404(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -309,26 +275,22 @@ namespace Web_Api.Tests.Controllers
 
             client.AddJwtToken(tokenJwt); // Add HTML header-request Authorization
 
-            var updateBookDto = new UpdateBookDto()
+            var updatePublisherDto = new UpdatePublisherDto()
             {
-                Id = 99999,
-                Name = "Test-book-EDITED",
-                Price = 11.11m,
-                IdPublisher = 1,
-                GenresId = new List<int> { 1 },
-                AuthorsId = new List<int> { 1 }
+                Id = 999999999,
+                Name = "Test-publisher-name",
             };
 
             // Act
-            var response = await client.PutAsJsonAsync(url, updateBookDto);
+            var response = await client.PutAsJsonAsync(url, updatePublisherDto);
 
             // Assert
             Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
         }
 
         [Theory]
-        [InlineData("api/book/2")]
-        public async Task BookDeleteAsyncTask_ReturnOk(string url)
+        [InlineData("api/publisher/3")]
+        public async Task PublisherDeleteAsyncTask_Return_Ok(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -345,8 +307,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book/3")]
-        public async Task BookDeleteAsyncTask_Return_Unauthorized_401(string url)
+        [InlineData("api/publisher/2")]
+        public async Task PublisherDeleteAsyncTask_Return_Unauthorized_401(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
@@ -359,8 +321,8 @@ namespace Web_Api.Tests.Controllers
         }
 
         [Theory]
-        [InlineData("api/book/2")]
-        public async Task BookDeleteAsyncTask_Return_Forbidden_403(string url)
+        [InlineData("api/publisher/2")]
+        public async Task PublisherDeleteAsyncTask_Return_Forbidden_403(string url)
         {
             // Arrange
             var client = _appFactory.CreateClient();
